+++ CHANGELOG.md
@@ -5,14 +5,19 @@
  and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
  
  ## [Unreleased]
-
+## [0.7.1] - 2020-05-05
+### Added
+ -redux icon and styles
+ ### Changed
+ -refactoring reducers
+  
 ## [0.7.0] - 2020-05-03
 ### Added
  - add yarn add redux And Proptypes
  - added model & icon project to about me
  - Create Components & redux skelenton For TodoList project example.
  - routing path /todolist 
- 
+ - new components
 ## [0.6.0] - 2020-04-29
 ### Added
  - added Molecule SkillsEducation view with provisional CSS and icon & first version Modals